requirements.txt
amqp==5.0.2
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==5.0.4
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Flask==1.1.2
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.3
pandas==1.1.5
prompt-toolkit==3.0.8
psycopg2-binary
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
rabbitmq==0.2.0
redis==3.5.3
six==1.15.0
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
flask_restful
